+++ NEWS.md
@@ -6,8 +6,8 @@ output: html_document
 
 This release is focused on adding new method `method="TMB"` which uses the package {marssTMB}.
 
-* Added methods "TMB", "BFGS.TMB" and "nlminb.TMB" and changed `MARSS()` to recognize these.
-* Added more descriptive methods for kem and optim: "EM.KFAS", "EM.KFSS", "BFGS.KFAS", "BFGS.KFSS"
+* Added methods "TMB", "BFGS_TMB" and "nlminb_TMB" and changed `MARSS()` to recognize these.
+* Added more descriptive methods for kem and optim: "EM_KFAS", "EM_KFSS", "BFGS_KFAS", "BFGS_KFSS"
 * Added the defaults to `R/onLoad.R` and made it easier to specify new methods and constraints on methods
 * Made changes to `checkMARSSinputs.R`, `is_marssMLE.R` and `is.validvarcov()` so they are not so specific to `method="BFGS"` but work for any methods with similar constraints.